+++ src/components/AntDesign/index.js
@@ -10,15 +10,15 @@ export { default as Table } from './Table/Table';
 export { default as Switch } from './Switch/Switch';
 export { default as Upload } from './Upload/index';
 export { default as Radio } from './Radio/Radio';
-export { default as Avatar } from './Avatar';
-export { default as ImgCrop } from './ImgCrop';
+export { default as Avatar } from './Avatar/Avatar';
+export { default as ImgCrop } from './ImgCrop/ImgCrop';
 export { default as Modal } from './Modal/Modal';
 export { default as PlusOutlined } from './Icons/PlusOutlined';
 export { default as LoadingOutlined } from './Icons/LoadingOutlined';
 export { default as DeleteOutlined } from './Icons/DeleteOutlined';
-export { default as Spin } from './Spin';
-export { default as Pagination } from './Pagination';
-export { default as AutoComplete } from './AutoComplete';
-export { default as Row } from './Row';
-export { default as Col } from './Col';
-export { default as Card } from './Card';
+export { default as Spin } from './Spin/Spin';
+export { default as Pagination } from './Pagination/Pagination';
+export { default as AutoComplete } from './AutoComplete/AutoComplete';
+export { default as Row } from './Row/Row';
+export { default as Col } from './Col/Col';
+export { default as Card } from './Card/Card';
